% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions-batch.R
\docType{class}
\name{AnalysisPipeline}
\alias{AnalysisPipeline}
\title{Class for constructing Analysis Pipelines for batch/ one-time analyeses}
\details{
Inherits the base class \link{BaseAnalysisPipeline} class which holds the metadata including the registry of available functions,
the data on which the pipeline is to be applied, as well as the pipeline itself

Additionally, this class is meant to be used for batch/ one-time processing. Contains additional slots to
hold the data frame to be used for the pipeline and associated schema

More details of how an object of this class should be initialized is provided in the
constructor - \link{initializeAnalysisPipeline}
}
\section{Slots}{

\describe{
\item{\code{input}}{The input dataset on which analysis is to be performed}

\item{\code{originalSchemaDf}}{Empty data frame representing the schema of the input}
}}

\seealso{
Other Package core functions for batch/one-time analyses: \code{\link{checkSchema}},
  \code{\link{generateReport}},
  \code{\link{initializeAnalysisPipeline}}
}
