% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisPipelines_package.R
\docType{package}
\name{analysisPipelines}
\alias{analysisPipelines}
\alias{analysisPipelines-package}
\title{analysisPipelines}
\description{
This package allows data scientists to compose and generate reports as a set of analytical operations.
The sequence of generation can be stored as a pipeline and reused, specifically for production systems where these
tasks are run repetitively. Additionally, the package implements a form of lazy evaluation where the pipelines are
 run on datasets only when outputs/ reports need to be generated
}
