% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisRecipes_package.R
\docType{package}
\name{analysisRecipes}
\alias{analysisRecipes}
\alias{analysisRecipes-package}
\title{analysisRecipes}
\description{
This package allows data scientists to compose and generate reports as a set of analytical operations.
The sequence of generation can be stored as recipes and reused, specifically for production systems where these
tasks are run repetitively. Additionally, the package implements a form of lazy evaluation where the recipes are
 run on datasets only when outputs/ reports need to be generated
}
