% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{class}
\name{BaseAnalysisPipeline}
\alias{BaseAnalysisPipeline}
\title{Base class for \code{AnalysisPipeline} and \code{StreamingAnalysisPipeline} objects}
\details{
The class which holds the metadata including the registry of available functions,
the data on which the pipeline is to be applied, as well as the pipeline itself, and serves
as the base class for various types of Pipeline objects such as Batch and Streaming.

This base class which contains the slots related to the registry, pipeline and output can be extended
to create custom class for specific scenarios if required.

The details of the constructor for this class can be found at \link{initializeBaseAnalysisPipeline}

In the documentation, objects of classes which are subclasses of this class are referred to as 'Pipeline' objects
}
\section{Slots}{

\describe{
\item{\code{pipeline}}{A tibble which holds functions to be called}

\item{\code{registry}}{A tibble which holds all the registered functions}

\item{\code{output}}{A list which holds all the functions output}
}}

\seealso{
Other Package core functions: \code{\link{assessEngineSetUp}},
  \code{\link{generateOutput}},
  \code{\link{genericPipelineException}},
  \code{\link{getInput}}, \code{\link{getOutputById}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initDfBasedOnType}},
  \code{\link{initializeBaseAnalysisPipeline}},
  \code{\link{loadPipeline}}, \code{\link{prepExecution}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{setInput}},
  \code{\link{updateObject}},
  \code{\link{visualizePipeline}}
}
