% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda-utilities.R
\name{CheckColumnType}
\alias{CheckColumnType}
\title{Check for type of column}
\usage{
CheckColumnType(dataVector)
}
\arguments{
\item{dataVector}{a data vector of a column}
}
\value{
column Type
}
\details{
Checking for type of columns in the datavector
}
\seealso{
Other Package EDA Utilites functions: \code{\link{bivarPlots}},
  \code{\link{correlationMatPlot}},
  \code{\link{getDatatype}}, \code{\link{ignoreCols}},
  \code{\link{multiVarOutlierPlot}},
  \code{\link{outlierPlot}},
  \code{\link{univarCatDistPlots}}
}
