% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{checkSchema}
\alias{checkSchema}
\title{Compare the schemas of two dataframes}
\usage{
checkSchema(dfOld, dfNew)
}
\arguments{
\item{dfOld}{Old dataframe}

\item{dfNew}{New dataframe}
}
\value{
Returns a list with details on added columns, removed columns, comparison between column classes, and a logical
        whether the schema has remained the same from the old dataframe to the new one
}
\details{
Compares the schemas of two dataframes, providing information on added and removed columns in the new dataframe
      as compared to the old
}
\seealso{
Other Package core functions: \code{\link{AnalysisPipeline}},
  \code{\link{assessEngineSetUp}},
  \code{\link{checkSchemaMatch}},
  \code{\link{generateOutput}},
  \code{\link{generateReport}},
  \code{\link{genericPipelineException}},
  \code{\link{getInput}}, \code{\link{getOuputByOrderId}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initialize}}, \code{\link{loadPipeline}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{setInput}},
  \code{\link{updateObject}}
}
\keyword{internal}
