% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R, R/core-functions-batch.R
\docType{methods}
\name{checkSchemaMatch}
\alias{checkSchemaMatch}
\alias{checkSchemaMatch,AnalysisPipeline-method}
\title{Checks the schema of the input to a Pipeline object against the original}
\usage{
checkSchemaMatch(object, newData)

\S4method{checkSchemaMatch}{AnalysisPipeline}(object, newData)
}
\arguments{
\item{object}{A Pipeline object}

\item{newData}{The newData that the pipeline is to be initialized with}
}
\value{
Returns a list with details on added columns, removed columns, comparison between column classes, and a logical
        whether the schema has remained the same from the old dataframe to the new one
}
\details{
Checks the schema of the new data frame that the pipeline is to be initialized with against
         the original schema that the pipeline was saved with. Provides a detailed comparison
}
\seealso{
Other Package core functions for batch/one-time analyses: \code{\link{AnalysisPipeline}},
  \code{\link{checkSchema}}, \code{\link{generateReport}},
  \code{\link{initializeAnalysisPipeline}}
}
