% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{computeEdges}
\alias{computeEdges}
\title{Computes edges (dependencies) in a pipeline given the joined tibble of the pipeline and registry}
\usage{
computeEdges(pipelineRegistryJoin)
}
\keyword{internal}
