% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-batch-eda-utilities.R
\name{correlationMatPlot}
\alias{correlationMatPlot}
\title{Correlation Matrix Plot}
\usage{
correlationMatPlot(dataset, methodused = "everything")
}
\arguments{
\item{dataset}{the dataset that needs to be loaded}

\item{methodused}{methods to be used for computing correlation}
}
\value{
Correlation Matrix graph
}
\description{
A correlation matrix is created and plotted across all the columns in the dataset
}
\seealso{
Other Package EDA Utilites functions: \code{\link{CheckColumnType}},
  \code{\link{bivarPlots}}, \code{\link{getDatatype}},
  \code{\link{ignoreCols}},
  \code{\link{multiVarOutlierPlot}},
  \code{\link{outlierPlot}},
  \code{\link{univarCatDistPlots}}
}
\concept{Package EDA Utilites functions}
