% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{data}
\name{.analysisPipelinesEnvir}
\alias{.analysisPipelinesEnvir}
\title{This section defines the environment which the package uses for maintaining the registry and an outputCache}
\format{An object of class \code{environment} of length 2.}
\usage{
.analysisPipelinesEnvir
}
\description{
This section defines the environment which the package uses for maintaining the registry and an outputCache
}
\keyword{internal}
