% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{.getCache}
\alias{.getCache}
\title{This is an internal function which returns the cache from the package namespace}
\usage{
.getCache()
}
\description{
This is an internal function which returns the cache from the package namespace
}
\keyword{internal}
