% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions-meta-pipelines.R
\name{.saveMetaPipeline}
\alias{.saveMetaPipeline}
\title{A method definition for saving meta-pipelines, called when the 'savePipeline' method is called against the
\code{MetaAnalysisPipeline} signature}
\usage{
.saveMetaPipeline(object, path)
}
\description{
A method definition for saving meta-pipelines, called when the 'savePipeline' method is called against the
\code{MetaAnalysisPipeline} signature
}
\keyword{internal}
