% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{.setRegistry}
\alias{.setRegistry}
\title{Internal function used to set the registry object in case of loading pipelines or meta-pipelines}
\usage{
.setRegistry(.registry)
}
\description{
Internal function used to set the registry object in case of loading pipelines or meta-pipelines
}
\keyword{internal}
