% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{.updateRegistry}
\alias{.updateRegistry}
\title{This is an internal function used to update the registry, in order to override existing function registrations}
\usage{
.updateRegistry(functionName, heading = "", engine = "r",
  exceptionHandlingFunction = as.character(substitute(genericPipelineException)),
  userDefined = F, isDataFunction = T, firstArgClass = "")
}
\description{
This is an internal function used to update the registry, in order to override existing function registrations
}
\keyword{internal}
