% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions-meta-pipelines.R
\name{.visualizeMetaPipeline}
\alias{.visualizeMetaPipeline}
\title{A method definition for visualizing meta-pipelines, called when the 'visualizePipeline' method is called against the
\code{MetaAnalysisPipeline} signature}
\usage{
.visualizeMetaPipeline(object)
}
\description{
A method definition for visualizing meta-pipelines, called when the 'visualizePipeline' method is called against the
\code{MetaAnalysisPipeline} signature
}
\keyword{internal}
