% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{generateOutput}
\alias{generateOutput}
\title{Generate a list of outputs from an \code{AnalysisPipeline} or \code{StreamingAnalysisPipeline} object}
\usage{
generateOutput(object)
}
\arguments{
\item{object}{object that contains input, pipeline, registry and output}
}
\value{
A list of the outputs in the sequence in which the pipeline was created
}
\details{
The sequence of operations stored in the\code{AnalysisPipeline} or \code{StreamingAnalysisPipeline} object
      are run and outputs generated, stored in a list
}
\seealso{
Other Package core functions: \code{\link{AnalysisPipeline}},
  \code{\link{assessEngineSetUp}},
  \code{\link{generateReport}}, \code{\link{getInput}},
  \code{\link{getOuputByOrderId}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initialize}}, \code{\link{loadPipeline}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{updateObject}}
}
