% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{getEndPoints}
\alias{getEndPoints}
\title{Obtains end nodes in a graph given nodes and edges}
\usage{
getEndPoints(nodes, edgeDf)
}
\description{
Obtains end nodes in a graph given nodes and edges
}
\keyword{internal}
