% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{getLoggerDetails}
\alias{getLoggerDetails}
\title{Obtains the logger configuration for the pipeline}
\usage{
getLoggerDetails(object)
}
\arguments{
\item{object}{A Pipeline object}
}
\value{
Logger configuration as a list
}
\details{
This function obtains the logger configuration for the pipeline.
}
\seealso{
Other Package core functions: \code{\link{BaseAnalysisPipeline}},
  \code{\link{assessEngineSetUp}},
  \code{\link{checkSchemaMatch}},
  \code{\link{generateOutput}},
  \code{\link{genericPipelineException}},
  \code{\link{getInput}}, \code{\link{getOutputById}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initDfBasedOnType}},
  \code{\link{initializeBaseAnalysisPipeline}},
  \code{\link{loadPipeline}},
  \code{\link{loadPredefinedFunctionRegistry}},
  \code{\link{prepExecution}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{setInput}},
  \code{\link{setLoggerDetails}},
  \code{\link{updateObject}},
  \code{\link{visualizePipeline}}
}
