% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{getOuputByOrderId}
\alias{getOuputByOrderId}
\title{Obtains a specific output}
\usage{
getOuputByOrderId(object, position)
}
\arguments{
\item{object}{The \code{AnalysisRecipe} or \code{SparkAnalysisRecipe} object}

\item{position}{The position of the function for which the output is desired in the sequence of operations in the recipe.}
}
\value{
List containing to elements
        - call: tibble with 1 row containing the function call for the output desired
        - output: output generated
}
\details{
Obtains a specific output from the \code{AnalysisRecipe} or \code{SparkAnalysisRecipe} object by passing the position
     of the function for which the output is desired, in the sequence of operations in the recipe. This can be obtained by passing the number
     under the 'order' column in the recipe table corresponding to the required function
}
\seealso{
Other Package core functions: \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{getInput}},
  \code{\link{getRecipe}}, \code{\link{getRegistry}},
  \code{\link{initialize}}, \code{\link{loadRecipe}},
  \code{\link{readInput}}, \code{\link{registerFunction}},
  \code{\link{saveRecipe}}, \code{\link{updateObject}}
}
