% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{getRecipe}
\alias{getRecipe}
\title{Obtain the recipe}
\usage{
getRecipe(object)
}
\arguments{
\item{object}{The \code{AnalysisRecipe} or \code{SparkAnalysisRecipe}  object}
}
\value{
Tibble describing the recipe
}
\details{
Obtains the recipe from the \code{AnalysisRecipe} or \code{SparkAnalysisRecipe} object as a tibble
}
\seealso{
Other Package core functions: \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{getInput}},
  \code{\link{getOuputByOrderId}},
  \code{\link{getRegistry}}, \code{\link{initialize}},
  \code{\link{loadRecipe}}, \code{\link{readInput}},
  \code{\link{registerFunction}}, \code{\link{saveRecipe}},
  \code{\link{updateObject}}
}
