% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{getRegistry}
\alias{getRegistry}
\title{Obtains the function registry}
\usage{
getRegistry(object)
}
\arguments{
\item{object}{The \code{AnalysisRecipe} or \code{SparkAnalysisRecipe}  object}
}
\value{
Tibble describing the registry
}
\details{
Obtains the function registry from the \code{AnalysisRecipe} or \code{SparkAnalysisRecipe} object as a tibble,
     including both predefined and user defined functions
}
\seealso{
Other Package core functions: \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{getInput}},
  \code{\link{getOuputByOrderId}}, \code{\link{getRecipe}},
  \code{\link{initialize}}, \code{\link{loadRecipe}},
  \code{\link{readInput}}, \code{\link{registerFunction}},
  \code{\link{saveRecipe}}, \code{\link{updateObject}}
}
