% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{getResponse}
\alias{getResponse}
\title{Obtains the response term from the formula}
\usage{
getResponse(f)
}
\arguments{
\item{f}{formula from which term is to be extracted.}
}
\value{
The response variable in the formula as a string
}
\description{
Obtains the response term from the formula
}
\details{
This is a helper function to extract the response variable from a formula
}
\examples{
library(analysisPipelines)
getResponse(y ~ x1 + x2)
}
