% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{getTerm}
\alias{getTerm}
\title{Obtains the dependency term from the formula}
\usage{
getTerm(f)
}
\arguments{
\item{f}{formula from which term is to be extracted.}
}
\value{
String with the terms
}
\description{
Obtains the dependency term from the formula
}
\details{
This is a helper function to extract the terms from a formula
}
\examples{
library(analysisPipelines)
getTerm(y ~ x)
}
