% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{getUpstreamDependencies}
\alias{getUpstreamDependencies}
\title{Obtains upstream dependencies for \code{AnalysisPipeline} objects}
\usage{
getUpstreamDependencies(row)
}
\description{
Obtains upstream dependencies for \code{AnalysisPipeline} objects
}
\keyword{internal}
