% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{identifyTopLevelRecursively}
\alias{identifyTopLevelRecursively}
\title{Recursive function to identify the toplogical levels of the functions in a pipeline}
\usage{
identifyTopLevelRecursively(input = list(topDf = dplyr::tibble(), nodes =
  c(), edgeDf = dplyr::tibble(), level = 1))
}
\description{
Recursive function to identify the toplogical levels of the functions in a pipeline
}
\keyword{internal}
