% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{identifyTopologicalLevels}
\alias{identifyTopologicalLevels}
\title{Identifies the topological levels of the functions in a pipeline}
\usage{
identifyTopologicalLevels(nodes = c(), edgeDf = dplyr::tibble(),
  topDf = dplyr::tibble(id = character(), level = character()),
  level = 1)
}
\description{
Identifies the topological levels of the functions in a pipeline
}
\keyword{internal}
