% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-batch-eda-utilities.R
\name{ignoreCols}
\alias{ignoreCols}
\title{Ignores the columns in the loaded dataframe object}
\usage{
ignoreCols(data, columns)
}
\arguments{
\item{data}{the dataframe object that needs to be loaded}

\item{columns}{the names of columns to be ignored from dataframe object}
}
\value{
Updated dataframe object
}
\description{
Ignores the columns in the loaded dataframe object
}
\details{
The columns selected are removed from the object
}
\examples{
ignoreCols(data = iris, columns = "Species")
}
\seealso{
Other Package EDA Utilites functions: \code{\link{CheckColumnType}},
  \code{\link{bivarPlots}},
  \code{\link{correlationMatPlot}},
  \code{\link{getDatatype}},
  \code{\link{multiVarOutlierPlot}},
  \code{\link{outlierPlot}},
  \code{\link{univarCatDistPlots}}
}
\concept{Package EDA Utilites functions}
