% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{initDfBasedOnType}
\alias{initDfBasedOnType}
\title{initializes the \code{AnalysisPipeline} object with the input based on the provided type}
\usage{
initDfBasedOnType(input, filePath)
}
\arguments{
\item{input}{Input dataframe}

\item{filePath}{File path where the .csv file is stored}
}
\value{
\code{AnalysisPipeline} object initialized with input
}
\details{
Transforms provided inputs into R data frame regardless of the input provided, be it Spark DataFrames
     or Python data frames
}
\seealso{
Other Package core functions: \code{\link{BaseAnalysisPipeline}},
  \code{\link{assessEngineSetUp}},
  \code{\link{generateOutput}},
  \code{\link{genericPipelineException}},
  \code{\link{getInput}}, \code{\link{getOutputById}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initializeBaseAnalysisPipeline}},
  \code{\link{loadPipeline}}, \code{\link{prepExecution}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{setInput}},
  \code{\link{updateObject}},
  \code{\link{visualizePipeline}}
}
\keyword{internal}
