% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-streaming-functions.R, R/core-functions.R
\docType{methods}
\name{initialize}
\alias{initialize}
\alias{initialize}
\title{Constructor for the \code{StreamingAnalysisPipeline} object}
\usage{
\S4method{initialize}{StreamingAnalysisPipeline}(.Object, input)

\S4method{initialize}{AnalysisPipeline}(.Object, input = data.frame(),
  filePath = "", workingInput = data.frame())
}
\arguments{
\item{.Object}{The \code{StreamingAnalysisPipeline} object}

\item{input}{The Spark DataFrame on which operations need to be performed}

\item{filePath}{File path for a .csv file to directly read in the dataset from}

\item{.Object}{The \code{AnalysisPipeline} object}

\item{input}{The data frame on which operations need to be performed}
}
\value{
an object of class "\code{StreamingAnalysisPipeline}", initialized with the input Spark DataFrame provided

an object of class "\code{AnalysisPipeline}", initialized with the input data frame provided
}
\details{
\code{input} needs to be provded and the argument needs to be of class \code{SparkDataFrame}, which is
     generally created through operations using SparkR

Either one of \code{input} or \code{filePath} need to be provided i.e. either the
     data frame or the file path to a csv file
}
\seealso{
Other Package core functions for Streaming Analyses: \code{\link{StreamingAnalysisPipeline}},
  \code{\link{loadStreamingPipeline}},
  \code{\link{registerStreamingFunction}}

Other Package core functions: \code{\link{AnalysisPipeline}},
  \code{\link{assessEngineSetUp}},
  \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{getInput}},
  \code{\link{getOuputByOrderId}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{loadPipeline}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{updateObject}}
}
