% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions-batch.R
\docType{methods}
\name{initializeAnalysisPipeline}
\alias{initializeAnalysisPipeline}
\title{Constructor for the \link{AnalysisPipeline} class}
\usage{
\S4method{initialize}{AnalysisPipeline}(.Object, ...,
  input = data.frame(), filePath = "")
}
\arguments{
\item{.Object}{The \code{AnalysisPipeline} object}

\item{input}{The data frame on which operations need to be performed}

\item{filePath}{File path for a .csv file to directly read in the dataset from}
}
\value{
an object of class \code{AnalysisPipeline}, initialized with the input data frame provided
}
\description{
Constructor for the \link{AnalysisPipeline} class
}
\details{
Either one of \code{input} or \code{filePath} need to be provided i.e. either the
     data frame or the file path to a csv file
}
\seealso{
Other Package core functions for batch/one-time analyses: \code{\link{AnalysisPipeline}},
  \code{\link{checkSchema}}, \code{\link{generateReport}}
}
\concept{Package core functions for batch/one-time analyses}
