% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{initializeLoggers}
\alias{initializeLoggers}
\title{intializes the loggers with the required appenders and layout based on the provided configuration}
\usage{
initializeLoggers(object)
}
\keyword{internal}
