% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-streaming-functions.R
\docType{methods}
\name{initializeStreamingAnalysisPipeline}
\alias{initializeStreamingAnalysisPipeline}
\title{Constructor for the \code{StreamingAnalysisPipeline} object}
\usage{
\S4method{initialize}{StreamingAnalysisPipeline}(.Object, input)
}
\arguments{
\item{.Object}{The \code{StreamingAnalysisPipeline} object}

\item{input}{The Spark DataFrame on which operations need to be performed}
}
\value{
an object of class "\code{StreamingAnalysisPipeline}", initialized with the input Spark DataFrame provided
}
\details{
\code{input} needs to be provded and the argument needs to be of class \code{SparkDataFrame}, which is
     generally created through operations using SparkR
}
\seealso{
Other Package core functions for Streaming Analyses: \code{\link{StreamingAnalysisPipeline}}
}
