% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{isDependencyParam}
\alias{isDependencyParam}
\title{Checks if the parameter is the dependency parameter}
\usage{
isDependencyParam(f)
}
\description{
Checks if the parameter is the dependency parameter
}
\keyword{internal}
