% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions-spark.R
\name{loadPipelineSpark}
\alias{loadPipelineSpark}
\title{Loads the \code{SparkAnalysisPipeline} object from the file system}
\usage{
loadPipelineSpark(RDSPath, input)
}
\arguments{
\item{RDSPath}{the path at which the .RDS file containing the pipeline is located}

\item{input}{Spark DataFrame with which the pipeline object should be initialized}
}
\value{
An \code{SparkAnalysisPipeline} object, optinally initialized with the data frame provided
}
\description{
### TODO
 @name loadPipelineSpark
}
\details{
The \code{SparkAnalysisPipeline} object is loaded into the file system from the file system
      based on the path specified.
}
\seealso{
Other Package core functions for Spark: \code{\link{initialize}},
  \code{\link{readInputSpark}},
  \code{\link{registerFunctionSpark}}
}
