% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{loadPredefinedFunctionRegistry}
\alias{loadPredefinedFunctionRegistry}
\title{Loading the registry of predefined functions}
\usage{
loadPredefinedFunctionRegistry()
}
\description{
Loading the registry of predefined functions
}
\details{
Loads the registry of predefined functions
}
\examples{
\dontrun{
  library(analysisPipelines)
  loadPredefinedFunctionRegistry()
}
}
\seealso{
Other Package core functions: \code{\link{BaseAnalysisPipeline-class}},
  \code{\link{MetaAnalysisPipeline-class}},
  \code{\link{assessEngineSetUp}},
  \code{\link{checkSchemaMatch}},
  \code{\link{createPipelineInstance}},
  \code{\link{exportAsMetaPipeline}},
  \code{\link{generateOutput}},
  \code{\link{genericPipelineException}},
  \code{\link{getInput}}, \code{\link{getLoggerDetails}},
  \code{\link{getOutputById}},
  \code{\link{getPipelinePrototype}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initDfBasedOnType}},
  \code{\link{initialize,BaseAnalysisPipeline-method}},
  \code{\link{loadMetaPipeline}},
  \code{\link{loadPipeline}}, \code{\link{loadRegistry}},
  \code{\link{prepExecution}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{saveRegistry}},
  \code{\link{setInput}}, \code{\link{setLoggerDetails}},
  \code{\link{updateObject}},
  \code{\link{visualizePipeline}}
}
\concept{Package core functions}
