% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions-spark.R
\name{loadRecipeSpark}
\alias{loadRecipeSpark}
\title{Loads the \code{SparkAnalysisRecipe} object from the file system}
\usage{
loadRecipeSpark(RDSPath, input)
}
\arguments{
\item{RDSPath}{the path at which the .RDS file containing the recipe is located}

\item{input}{Spark DataFrame with which the recipe object should be initialized}
}
\value{
An \code{SparkAnalysisRecipe} object, optinally initialized with the data frame provided
}
\description{
### TODO
 @name loadRecipeSpark
}
\details{
The \code{SparkAnalysisRecipe} object is loaded into the file system from the file system
      based on the path specified.
}
\seealso{
Other Package core functions for Spark: \code{\link{initialize}},
  \code{\link{readInputSpark}},
  \code{\link{registerFunctionSpark}}
}
