% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-streaming-functions.R
\name{loadStreamingPipeline}
\alias{loadStreamingPipeline}
\title{Loads the \code{StreamingAnalysisPipeline} object from the file system}
\usage{
loadStreamingPipeline(RDSPath, input)
}
\arguments{
\item{RDSPath}{the path at which the .RDS file containing the pipeline is located}

\item{input}{Spark Sructuded Streaming DataFrame with which the pipeline object should be initialized}
}
\value{
An \code{StreamingAnalysisPipeline} object, optinally initialized with the data frame provided
}
\description{
### TODO
 @name loadStreamingPipeline
}
\details{
The \code{StreamingAnalysisPipeline} object is loaded into the file system from the file system
      based on the path specified.
}
\seealso{
Other Package core functions for Streaming Analyses: \code{\link{initialize}},
  \code{\link{readStreamingInput}},
  \code{\link{registerStreamingFunction}}
}
