% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{prepExecution}
\alias{prepExecution}
\title{Prepare the pipleline for execution}
\usage{
prepExecution(object)
}
\value{
Updated \code{AnalysisPipeline} \code{StreamingAnalysisPipeline} object
}
\details{
The pipeline is prepared for execution by identifying the graph of the pipeline as well as its topological ordering,
and dependency map in order to prepare for execution
}
\seealso{
Other Package core functions: \code{\link{BaseAnalysisPipeline}},
  \code{\link{assessEngineSetUp}},
  \code{\link{checkSchemaMatch}},
  \code{\link{generateOutput}},
  \code{\link{genericPipelineException}},
  \code{\link{getInput}}, \code{\link{getLoggerDetails}},
  \code{\link{getOutputById}}, \code{\link{getPipeline}},
  \code{\link{getRegistry}},
  \code{\link{initDfBasedOnType}},
  \code{\link{initializeBaseAnalysisPipeline}},
  \code{\link{loadPipeline}},
  \code{\link{loadPredefinedFunctionRegistry}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{setInput}},
  \code{\link{setLoggerDetails}},
  \code{\link{updateObject}},
  \code{\link{visualizePipeline}}
}
