% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{saveRecipe}
\alias{saveRecipe}
\title{Saves the \code{AnalysisRecipe} object to the file system}
\usage{
saveRecipe(object, RDSPath)
}
\arguments{
\item{object}{object that contains input, recipe, registry and output}

\item{RDSPath}{the path at which the .RDS file containing the recipe should be stored}
}
\value{
Does not return a value
}
\details{
The \code{AnalysisRecipe} object is saved to the file system in the paths specified
}
\seealso{
Other Package core functions: \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{initialize}},
  \code{\link{loadRecipe}}, \code{\link{readInputSpark}},
  \code{\link{readInput}},
  \code{\link{registerFunctionSpark}},
  \code{\link{registerFunction}},
  \code{\link{sparkRSessionCreateIfNotPresent}},
  \code{\link{updateObject}}
}
