% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{saveRegistry}
\alias{saveRegistry}
\title{Saves the registry to the file system}
\usage{
saveRegistry(path)
}
\arguments{
\item{path}{path on the file system, where the registry is to be saved to}
}
\description{
Saves the registry to the file system
}
\details{
This function saves  the existing function registry and associated function definition loaded in the
environment into a file.
}
\examples{
\dontrun{
  library(analysisPipelines)
  saveRegistry(path = "./registry.RDS")
}
}
\seealso{
Other Package core functions: \code{\link{BaseAnalysisPipeline-class}},
  \code{\link{MetaAnalysisPipeline-class}},
  \code{\link{assessEngineSetUp}},
  \code{\link{checkSchemaMatch}},
  \code{\link{createPipelineInstance}},
  \code{\link{exportAsMetaPipeline}},
  \code{\link{generateOutput}},
  \code{\link{genericPipelineException}},
  \code{\link{getInput}}, \code{\link{getLoggerDetails}},
  \code{\link{getOutputById}},
  \code{\link{getPipelinePrototype}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initDfBasedOnType}},
  \code{\link{initialize,BaseAnalysisPipeline-method}},
  \code{\link{loadMetaPipeline}},
  \code{\link{loadPipeline}},
  \code{\link{loadPredefinedFunctionRegistry}},
  \code{\link{loadRegistry}}, \code{\link{prepExecution}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{setInput}},
  \code{\link{setLoggerDetails}},
  \code{\link{updateObject}},
  \code{\link{visualizePipeline}}
}
\concept{Package core functions}
