% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-helper-utilites-python.R
\name{setPythonEnvir}
\alias{setPythonEnvir}
\title{Sets the python environment to be used}
\usage{
setPythonEnvir(type = "conda", pathOrEnvirName = "base")
}
\arguments{
\item{type}{Type of python environment. Takes three possible vales - 'conda' for Anaconda environments,
'virtualenv' for Virtual environments, and 'python' to manually set the python path to use}

\item{pathOrEnvirName}{Name of the environment for Anaconda and Virtual environments,
or the Python path when type is 'python'}
}
\description{
Sets the python environment to be used
}
\details{
Wrapper function over reticulate functions to set a python environment to be used
}
