% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{setUpstreamDependencies}
\alias{setUpstreamDependencies}
\title{Sets upstream dependencies for the entire pipeline}
\usage{
setUpstreamDependencies(pipeline)
}
\description{
Sets upstream dependencies for the entire pipeline
}
\keyword{internal}
