% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-batch-eda-utilities.R
\name{univarCatDistPlots}
\alias{univarCatDistPlots}
\title{Univariate Categoric Distribution}
\usage{
univarCatDistPlots(data, uniCol, priColor = "blue", optionalPlots = 0)
}
\arguments{
\item{data}{the dataset where the column on which the plot is to be generated is present}

\item{uniCol}{the name of column on which the plot needs to be generated}

\item{priColor}{the primary color for the plots}

\item{optionalPlots}{A Flag for optional plots}
}
\value{
A univariate categoric distribution plot
}
\description{
Univariate Categoric Distribution
}
\details{
A univariate distribution graph on the selected categorical columns from the dataframe
}
\examples{
univarCatDistPlots(data = iris, uniCol = "Species")
}
\seealso{
Other Package EDA Utilites functions: \code{\link{CheckColumnType}},
  \code{\link{bivarPlots}},
  \code{\link{correlationMatPlot}},
  \code{\link{getDatatype}}, \code{\link{ignoreCols}},
  \code{\link{multiVarOutlierPlot}},
  \code{\link{outlierPlot}}
}
\concept{Package EDA Utilites functions}
