% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{updateObject}
\alias{updateObject}
\title{Update the \code{AnalysisPipeline} or \code{StreamingAnalysisPipeline} object by adding an operation to the pipeline}
\usage{
updateObject(object, operation, heading = "", parameters, outAsIn = F)
}
\arguments{
\item{object}{object that contains input, pipeline, registry and output}

\item{operation}{function name to be updated in tibble}

\item{heading}{heading of that section in report}

\item{parameters}{parameters passed to that function}

\item{outAsIn}{whether to use original input or output from previous function}
}
\value{
Updated \code{AnalysisPipeline} object
}
\details{
The specified operation along with the heading and parameters is updated in the pipeline slot
      of the  \code{AnalysisPipeline} or \code{StreamingAnalysisPipeline} object, where the sequence of operations
     to be performed is stored
}
\seealso{
Other Package core functions: \code{\link{AnalysisPipeline}},
  \code{\link{assessEngineSetUp}},
  \code{\link{checkSchemaMatch}},
  \code{\link{checkSchema}}, \code{\link{generateOutput}},
  \code{\link{generateReport}},
  \code{\link{genericPipelineException}},
  \code{\link{getInput}}, \code{\link{getOuputByOrderId}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initialize}}, \code{\link{loadPipeline}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{setInput}}
}
