% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{updatePackageRegistry}
\alias{updatePackageRegistry}
\title{Updates the package registry}
\usage{
updatePackageRegistry(functionName, functionHeader, flag)
}
\arguments{
\item{functionName}{the name of the function}

\item{functionHeader}{the header caption that will feature in the report for this function's output}

\item{flag}{a boolean which dictates if the 'functionName' function returns a data.frame to be used an input}
}
\value{
An \code{AnalysisRecipe} object, optinally initialized with the data frame provided
}
\details{
Updates the registry of predefined functions available in the package (For developer use)
}
\seealso{
Other Package core functions: \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{initialize}},
  \code{\link{loadRecipe}}, \code{\link{readInput}},
  \code{\link{registerFunction}}, \code{\link{saveRecipe}},
  \code{\link{updateObject}}
}
\keyword{internal}
