% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisPipelines_package.R
\docType{package}
\name{analysisPipelines}
\alias{analysisPipelines}
\alias{analysisPipelines-package}
\title{analysisPipelines}
\description{
The package aims at enabling data scientists to compose pipelines of analysis which consist of data manipulation,
exploratory analysis & reporting, as well as modeling steps. It also aims to enable data scientists to use tools
of their choice through an R interface, and compose interoperable pipelines between R, Spark, and Python.
}
\details{
Important Note - This package uses 'SparkR' to interact with Spark and automatically installs it if not present
from a Github repo as 'SparkR' is not distrubuted on CRAN
}
