% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{assessEngineSetUp}
\alias{assessEngineSetUp}
\title{Assesses engine (R, Spark, Python) set up}
\usage{
assessEngineSetUp(object)
}
\arguments{
\item{object}{object that contains input, pipeline, registry and output}
}
\value{
Tibble containing the details of available engines, whether they are required for a recipe, a logical reporting
        whether the engine has been set up, and comments.
}
\details{
Assesses whether engines required for executing functions in an \code{AnalysisPipeline} object have been set up
}
\seealso{
Other Package core functions: \code{\link{AnalysisPipeline}},
  \code{\link{checkSchemaMatch}},
  \code{\link{checkSchema}}, \code{\link{generateOutput}},
  \code{\link{generateReport}},
  \code{\link{genericPipelineException}},
  \code{\link{getInput}}, \code{\link{getOuputByOrderId}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initialize}}, \code{\link{loadPipeline}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{setInput}},
  \code{\link{updateObject}}
}
