% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{assessEngineSetUp}
\alias{assessEngineSetUp}
\title{Assesses engine (R, Spark, Python, Spark Structured Streaming) set up}
\usage{
assessEngineSetUp(object)
}
\arguments{
\item{object}{object that contains input, pipeline, registry and output}
}
\value{
Tibble containing the details of available engines, whether they are required for a recipe, a logical reporting
        whether the engine has been set up, and comments.
}
\details{
Assesses whether engines required for executing functions in an \code{AnalysisPipeline} or \code{StreamingAnalysisPipeline}
      object have been set up

This method is implemented on the base class as it is a shared functionality types of Analysis Pipelines
which extend this class
}
\seealso{
Other Package core functions: \code{\link{BaseAnalysisPipeline}},
  \code{\link{generateOutput}},
  \code{\link{genericPipelineException}},
  \code{\link{getInput}}, \code{\link{getOutputById}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initDfBasedOnType}},
  \code{\link{initializeBaseAnalysisPipeline}},
  \code{\link{loadPipeline}}, \code{\link{prepExecution}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{setInput}},
  \code{\link{updateObject}},
  \code{\link{visualizePipeline}}
}
