% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{checkSchemaMatch}
\alias{checkSchemaMatch}
\title{Checks the schema of a pipeline against the original}
\usage{
checkSchemaMatch(object, newData)
}
\arguments{
\item{object}{The \code{AnalysisPipeline} object}

\item{newData}{The newData that the pipeline is to be initialized with}
}
\value{
Returns a list with details on added columns, removed columns, comparison between column classes, and a logical
        whether the schema has remained the same from the old dataframe to the new one
}
\details{
Checks the schema of the new data frame that the pipeline is to be initialized with against
         the original schema that the pipeline was saved with. Provides a detailed comparison
}
\seealso{
Other Package core functions: \code{\link{AnalysisPipeline}},
  \code{\link{assessEngineSetUp}},
  \code{\link{checkSchema}}, \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{getInput}},
  \code{\link{getOuputByOrderId}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initialize}}, \code{\link{loadPipeline}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{setInput}},
  \code{\link{updateObject}}
}
