% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{generateOutput}
\alias{generateOutput}
\title{Generate a list of outputs from an \code{AnalysisRecipe} object}
\usage{
generateOutput(object)
}
\arguments{
\item{object}{object that contains input, recipe, registry and output}
}
\value{
A list of the outputs in the sequence in which the recipe was created
}
\details{
The sequence of operations stored in the \code{AnalysisRecipe} object are run and outputs generated,
      stored in a list
}
\seealso{
Other Package core functions: \code{\link{generateReport}},
  \code{\link{initialize}}, \code{\link{loadRecipe}},
  \code{\link{readInput}},
  \code{\link{registerFunctionSpark}},
  \code{\link{registerFunction}}, \code{\link{saveRecipe}},
  \code{\link{sparkRSessionCreateIfNotPresent}},
  \code{\link{updateObject}}
}
