% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{generateReport}
\alias{generateReport}
\title{Generate a HTML report from an \code{AnalysisRecipe} object}
\usage{
generateReport(object, path)
}
\arguments{
\item{object}{object that contains input, recipe, registry and output}

\item{path}{path on the file system, where the generated html report should be stored}
}
\value{
Updated \code{AnalysisRecipe} object
}
\details{
The sequence of operations stored in the \code{AnalysisRecipe} object are run, outputs generated,
      and a HTML report is generated with outputs in the same sequence as the pipeline created by the user
}
\seealso{
Other Package core functions: \code{\link{generateOutput}},
  \code{\link{getInput}}, \code{\link{getOuputByOrderId}},
  \code{\link{getRecipe}}, \code{\link{getRegistry}},
  \code{\link{initialize}}, \code{\link{loadRecipe}},
  \code{\link{readInput}}, \code{\link{registerFunction}},
  \code{\link{saveRecipe}}, \code{\link{updateObject}}
}
