% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-batch-eda-utilities.R
\name{getDatatype}
\alias{getDatatype}
\title{Get Data Type}
\usage{
getDatatype(dataset)
}
\arguments{
\item{dataset}{a dataset which needs to be loaded}
}
\value{
list with \code{numeric_cols} and \code{cat_cols}
}
\description{
Get Data Type
}
\details{
Based on the datatype the columns are seperated into categorical and numerical columns
}
\examples{
getDatatype(iris)
}
\seealso{
Other Package EDA Utilites functions: \code{\link{CheckColumnType}},
  \code{\link{bivarPlots}},
  \code{\link{correlationMatPlot}},
  \code{\link{ignoreCols}},
  \code{\link{multiVarOutlierPlot}},
  \code{\link{outlierPlot}},
  \code{\link{univarCatDistPlots}}
}
\concept{Package EDA Utilites functions}
