% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{getInput}
\alias{getInput}
\title{Obtains the initializedInput}
\usage{
getInput(object)
}
\arguments{
\item{object}{The \code{AnalysisRecipe} or \code{SparkAnalysisRecipe} object}
}
\value{
Dataframe for an \code{AnalysisRecipe} & SparkDataFrame for a \code{SparkAnalysisRecipe}
}
\details{
Obtains the input from the \code{AnalysisRecipe} or \code{SparkAnalysisRecipe} object
}
\seealso{
Other Package core functions: \code{\link{generateOutput}},
  \code{\link{generateReport}},
  \code{\link{getOuputByOrderId}}, \code{\link{getRecipe}},
  \code{\link{getRegistry}}, \code{\link{initialize}},
  \code{\link{loadRecipe}}, \code{\link{readInput}},
  \code{\link{registerFunction}}, \code{\link{saveRecipe}},
  \code{\link{updateObject}}
}
