% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{getInput}
\alias{getInput}
\title{Obtains the initializedInput}
\usage{
getInput(object)
}
\arguments{
\item{object}{The \code{AnalysisPipeline} or \code{StreamingAnalysisPipeline} object}
}
\value{
Dataframe for an \code{AnalysisPipeline} & SparkDataFrame for a \code{StreamingAnalysisPipeline}
}
\details{
Obtains the input from the \code{AnalysisPipeline} or \code{StreamingAnalysisPipeline} object

This method is implemented on the base class as it is a shared functionality types of Analysis Pipelines
which extend this class
}
\seealso{
Other Package core functions: \code{\link{BaseAnalysisPipeline}},
  \code{\link{assessEngineSetUp}},
  \code{\link{checkSchemaMatch}},
  \code{\link{generateOutput}},
  \code{\link{genericPipelineException}},
  \code{\link{getLoggerDetails}},
  \code{\link{getOutputById}}, \code{\link{getPipeline}},
  \code{\link{getRegistry}},
  \code{\link{initDfBasedOnType}},
  \code{\link{initializeBaseAnalysisPipeline}},
  \code{\link{loadPipeline}},
  \code{\link{loadPredefinedFunctionRegistry}},
  \code{\link{prepExecution}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{setInput}},
  \code{\link{setLoggerDetails}},
  \code{\link{updateObject}},
  \code{\link{visualizePipeline}}
}
