% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{getInput}
\alias{getInput}
\title{Obtains the initializedInput}
\usage{
getInput(object)
}
\arguments{
\item{object}{The \code{AnalysisPipeline} or \code{SparkAnalysisPipeline} object}
}
\value{
Dataframe for an \code{AnalysisPipeline} & SparkDataFrame for a \code{SparkAnalysisPipeline}
}
\details{
Obtains the input from the \code{AnalysisPipeline} or \code{SparkAnalysisPipeline} object
}
\seealso{
Other Package core functions: \code{\link{generateOutput}},
  \code{\link{generateReport}},
  \code{\link{getOuputByOrderId}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initialize}}, \code{\link{loadPipeline}},
  \code{\link{readInput}}, \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{updateObject}}
}
