% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{getOuputByOrderId}
\alias{getOuputByOrderId}
\title{Obtains a specific output}
\usage{
getOuputByOrderId(object, position, includeCall = F)
}
\arguments{
\item{object}{The \code{AnalysisPipeline} or \code{StreamingAnalysisPipeline} object}

\item{position}{The position of the function for which the output is desired in the sequence of operations in the pipeline.}

\item{includeCall}{Logical which defines whether the call used to generate the output should be returned. By, default this is false}
}
\value{
If includeCall = F, the output object generated by the function is returned

If includeCall = T, it is a list containing to elements
        - call: tibble with 1 row containing the function call for the output desired
        - output: output generated
}
\details{
Obtains a specific output from the \code{AnalysisPipeline} or \code{StreamingAnalysisPipeline} object by passing the position
     of the function for which the output is desired, in the sequence of operations in the pipeline. This can be obtained by passing the number
     under the 'order' column in the pipeline table corresponding to the required function
}
\seealso{
Other Package core functions: \code{\link{AnalysisPipeline}},
  \code{\link{assessEngineSetUp}},
  \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{getInput}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initialize}}, \code{\link{loadPipeline}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{updateObject}}
}
