% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{getPipeline}
\alias{getPipeline}
\title{Obtain the pipeline}
\usage{
getPipeline(object)
}
\arguments{
\item{object}{The \code{AnalysisPipeline} or \code{SparkAnalysisPipeline}  object}
}
\value{
Tibble describing the pipeline
}
\details{
Obtains the pipeline from the \code{AnalysisPipeline} or \code{SparkAnalysisPipeline} object as a tibble
}
\seealso{
Other Package core functions: \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{getInput}},
  \code{\link{getOuputByOrderId}},
  \code{\link{getRegistry}}, \code{\link{initialize}},
  \code{\link{loadPipeline}}, \code{\link{readInput}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{updateObject}}
}
