% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{getPipeline}
\alias{getPipeline}
\title{Obtain the pipeline}
\usage{
getPipeline(object)
}
\arguments{
\item{object}{The \code{AnalysisPipeline} or \code{StreamingAnalysisPipeline} object}
}
\value{
Tibble describing the pipeline
}
\details{
Obtains the pipeline from the \code{AnalysisPipeline} or \code{StreamingAnalysisPipeline} object as a tibble

This method is implemented on the base class as it is a shared functionality types of Analysis Pipelines
which extend this class
}
\seealso{
Other Package core functions: \code{\link{BaseAnalysisPipeline}},
  \code{\link{assessEngineSetUp}},
  \code{\link{generateOutput}},
  \code{\link{genericPipelineException}},
  \code{\link{getInput}}, \code{\link{getOuputByOrderId}},
  \code{\link{getRegistry}},
  \code{\link{initDfBasedOnType}},
  \code{\link{initializeBaseAnalysisPipeline}},
  \code{\link{loadPipeline}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{setInput}},
  \code{\link{updateObject}}
}
