% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{getRegistry}
\alias{getRegistry}
\title{Obtains the function registry}
\usage{
getRegistry(object)
}
\arguments{
\item{object}{The \code{AnalysisPipeline} or \code{StreamingAnalysisPipeline}  object}
}
\value{
Tibble describing the registry
}
\details{
Obtains the function registry from the \code{AnalysisPipeline} or \code{StreamingAnalysisPipeline} object as a tibble,
     including both predefined and user defined functions
}
\seealso{
Other Package core functions: \code{\link{assessEngineSetUp}},
  \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{getInput}},
  \code{\link{getOuputByOrderId}},
  \code{\link{getPipeline}}, \code{\link{initialize}},
  \code{\link{loadPipeline}}, \code{\link{readInput}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{updateObject}}
}
