% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions-spark.R, R/core-functions.R
\docType{methods}
\name{initialize}
\alias{initialize}
\alias{initialize}
\title{Constructor for the \code{SparkAnalysisRecipe} object}
\usage{
\S4method{initialize}{SparkAnalysisRecipe}(.Object, input)

\S4method{initialize}{AnalysisRecipe}(.Object, input = data.frame(),
  filePath = "")
}
\arguments{
\item{.Object}{The \code{SparkAnalysisRecipe} object}

\item{input}{The data frame on which operations need to be performed}

\item{filePath}{File path for a .csv file to directly read in the dataset from}

\item{.Object}{The \code{AnalysisRecipe} object}

\item{input}{The data frame on which operations need to be performed}

\item{filePath}{File path for a .csv file to directly read in the dataset from}
}
\value{
an object of class "\code{SparkAnalysisRecipe}", initialized with the input data frame provided

an object of class "\code{AnalysisRecipe}", initialized with the input data frame provided
}
\details{
Either one of \code{input} or \code{filePath} need to be provided i.e. either the
     data frame or the file path to a csv file

Either one of \code{input} or \code{filePath} need to be provided i.e. either the
     data frame or the file path to a csv file
}
\seealso{
Other Package core functions: \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{loadRecipe}},
  \code{\link{readInputSpark}}, \code{\link{readInput}},
  \code{\link{registerFunctionSpark}},
  \code{\link{registerFunction}}, \code{\link{saveRecipe}},
  \code{\link{sparkRSessionCreateIfNotPresent}},
  \code{\link{updateObject}}

Other Package core functions: \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{loadRecipe}},
  \code{\link{readInputSpark}}, \code{\link{readInput}},
  \code{\link{registerFunctionSpark}},
  \code{\link{registerFunction}}, \code{\link{saveRecipe}},
  \code{\link{sparkRSessionCreateIfNotPresent}},
  \code{\link{updateObject}}
}
