% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{methods}
\name{initializeBaseAnalysisPipeline}
\alias{initializeBaseAnalysisPipeline}
\title{This is the constructor for the \link{BaseAnalysisPipeline} class}
\usage{
\S4method{initialize}{BaseAnalysisPipeline}(.Object)
}
\arguments{
\item{.Object}{The \code{BaseAnalysisPipeline} object}
}
\value{
an object of class "\code{BaseAnalysisPipeline}"
}
\details{
This is a constructor function for the base class for various types of Analysis Pipelines. This method gets
     internally called by \code{AnalysisPipeline} and \code{StreamingAnalysisPipeline} constructors.
}
\seealso{
Other Package core functions: \code{\link{BaseAnalysisPipeline}},
  \code{\link{assessEngineSetUp}},
  \code{\link{generateOutput}},
  \code{\link{genericPipelineException}},
  \code{\link{getInput}}, \code{\link{getOutputById}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initDfBasedOnType}},
  \code{\link{loadPipeline}}, \code{\link{prepExecution}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{setInput}},
  \code{\link{updateObject}},
  \code{\link{visualizePipeline}}
}
