% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{isDependencyParam}
\alias{isDependencyParam}
\title{Checks if the parameter is the dependency parameter}
\usage{
isDependencyParam(f)
}
\arguments{
\item{f}{formula from which term is to be extracted.}
}
\value{
Logical as to whether it is a dependency parameter
}
\description{
Checks if the parameter is the dependency parameter
}
\details{
This is a helper function to check if the formula provided is a dependency parameter,
as per the package's formula semantics, capturing function dependencies
}
\examples{
library(analysisPipelines)
isDependencyParam(~f1)
}
