% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{loadPipeline}
\alias{loadPipeline}
\title{Loads the \code{AnalysisPipeline} object from the file system}
\usage{
loadPipeline(RDSPath, input = data.frame(), filePath = "")
}
\arguments{
\item{RDSPath}{the path at which the .RDS file containing the pipeline is located}

\item{input}{(optional) data frame with which the pipeline object should be initialized}

\item{filePath}{(optional) path where a dataset in .CSV format is present which is to be loaded}
}
\value{
An \code{AnalysisPipeline} object, optinally initialized with the data frame provided
}
\details{
The \code{AnalysisPipeline} object is loaded into the file system from the file system
      based on the path specified.

Optionally, the \code{input} parameter can be provided to
      initialize the \code{AnalysisPipeline} object with a data frame present in the R session.
      Another provided option, is to specify a filePath where the input dataset is present (in a .CSV format)
      and the object will be initialized with this data frame. The \code{filePath} parameter takes precedence over
      \code{input} parameter
}
\seealso{
Other Package core functions: \code{\link{assessEngineSetUp}},
  \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{getInput}},
  \code{\link{getOuputByOrderId}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initialize}}, \code{\link{readInput}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{updateObject}}
}
