% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{registerFunction}
\alias{registerFunction}
\title{Register a user-defined function to be used with \code{AnalysisPipeline} objects}
\usage{
registerFunction(object, functionName, heading = "", outAsIn = F,
  engine = "r", loadPipeline = F, userDefined = T, session = session)
}
\arguments{
\item{object}{object that contains input, pipeline, registry and output}

\item{functionName}{name of function to be registered}

\item{heading}{heading of that section in report}

\item{outAsIn}{whether to use original input or output from previous function}

\item{engine}{specifies which engine the function is to be run on. Available engines include "r", "spark", and "python"}

\item{loadPipeline}{logical parameter to see if function is being used in loadPipeline or not}

\item{session}{to load shiny session in the function}
}
\value{
Updated \code{AnalysisPipeline} object
}
\details{
The specified operation along with the heading and parameters is updated in the pipeline slot
      of the AnalysisPipeline object, where the sequence of operations to be performed is stored
}
\seealso{
Other Package core functions: \code{\link{assessEngineSetUp}},
  \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{getInput}},
  \code{\link{getOuputByOrderId}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initialize}}, \code{\link{loadPipeline}},
  \code{\link{readInput}}, \code{\link{savePipeline}},
  \code{\link{updateObject}}
}
