% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{registerFunction}
\alias{registerFunction}
\title{Register a user-defined function to be used with \code{AnalysisRecipe} objects}
\usage{
registerFunction(object, functionName, heading = "", outAsIn = F,
  loadRecipe = F, session = session)
}
\arguments{
\item{object}{object that contains input, recipe, registry and output}

\item{functionName}{name of function to be registered}

\item{heading}{heading of that section in report}

\item{outAsIn}{whether to use original input or output from previous function}

\item{loadRecipe}{logical parameter to see if function is being used in loadRecipe or not}

\item{session}{to load shiny session in the function}
}
\value{
Updated \code{AnalysisRecipe} object
}
\details{
The specified operation along with the heading and parameters is updated in the recipe slot
      of the AnalysisRecipe object, where the sequence of operations to be performed is stored
}
\seealso{
Other Package core functions: \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{initialize}},
  \code{\link{loadRecipe}}, \code{\link{readInput}},
  \code{\link{saveRecipe}}, \code{\link{updateObject}}
}
