% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions-spark.R
\name{registerFunctionSpark}
\alias{registerFunctionSpark}
\title{Register a user-defined function to be used with \code{SparkAnalysisRecipe} objects specifically for Spark
       DataFrames}
\usage{
registerFunctionSpark(object, functionName, heading = "", outAsIn = F,
  loadRecipe = F, session = session)
}
\arguments{
\item{object}{object that contains input, recipe, registry and output}

\item{functionName}{name of function to be registered}

\item{heading}{heading of that section in report}

\item{outAsIn}{whether to use output of this function as input to next}

\item{loadRecipe}{logical parameter to see if function is being used in loadRecipe or not}

\item{session}{to load shiny session in the function}
}
\value{
Updated \code{AnalysisRecipe} object
}
\details{
The specified operation along with the heading and parameters is updated in the recipe slot
      of the AnalysisRecipe object, where the sequence of operations to be performed is stored. This function
      is used to register functions which are designed to operate on Spark DataFrames
}
\seealso{
Other Package core functions: \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{initialize}},
  \code{\link{loadRecipe}}, \code{\link{readInputSpark}},
  \code{\link{readInput}}, \code{\link{registerFunction}},
  \code{\link{saveRecipe}},
  \code{\link{sparkRSessionCreateIfNotPresent}},
  \code{\link{updateObject}}
}
