% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{savePipeline}
\alias{savePipeline}
\title{Saves the \code{AnalysisPipeline} or \code{StreamingAnalysisPipeline}  object to the file system without outputs}
\usage{
savePipeline(object, RDSPath)
}
\arguments{
\item{object}{object that contains input, pipeline, registry and output}

\item{RDSPath}{the path at which the .RDS file containing the pipeline should be stored}
}
\value{
Does not return a value
}
\details{
The \code{AnalysisPipeline} object is saved to the file system in the paths specified
}
\seealso{
Other Package core functions: \code{\link{AnalysisPipeline}},
  \code{\link{assessEngineSetUp}},
  \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{getInput}},
  \code{\link{getOuputByOrderId}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initialize}}, \code{\link{loadPipeline}},
  \code{\link{registerFunction}},
  \code{\link{updateObject}}
}
