% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{saveRecipe}
\alias{saveRecipe}
\title{Saves the \code{AnalysisRecipe} or \code{SparkAnalysisRecipe}  object to the file system without outputs}
\usage{
saveRecipe(object, RDSPath)
}
\arguments{
\item{object}{object that contains input, recipe, registry and output}

\item{RDSPath}{the path at which the .RDS file containing the recipe should be stored}
}
\value{
Does not return a value
}
\details{
The \code{AnalysisRecipe} object is saved to the file system in the paths specified
}
\seealso{
Other Package core functions: \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{getInput}},
  \code{\link{getOuputByOrderId}}, \code{\link{getRecipe}},
  \code{\link{getRegistry}}, \code{\link{initialize}},
  \code{\link{loadRecipe}}, \code{\link{readInput}},
  \code{\link{registerFunction}},
  \code{\link{updateObject}}
}
