% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{setInput}
\alias{setInput}
\title{Sets the input for an \code{AnalysisPipeline} or \code{StreamingAnalysisPipeline} object}
\usage{
setInput(object, input, filePath = "")
}
\arguments{
\item{object}{object that contains input, pipeline, registry and output}

\item{input}{the input data frame}

\item{filePath}{path to the file which needs to be read (currently supports .csv files)}
}
\value{
Updated \code{AnalysisPipeline} object
}
\details{
Assigns the input to the pipeline for an  \code{AnalysisPipeline} or \code{StreamingAnalysisPipeline} object,
}
\seealso{
Other Package core functions: \code{\link{AnalysisPipeline}},
  \code{\link{assessEngineSetUp}},
  \code{\link{checkSchemaMatch}},
  \code{\link{checkSchema}}, \code{\link{generateOutput}},
  \code{\link{generateReport}},
  \code{\link{genericPipelineException}},
  \code{\link{getInput}}, \code{\link{getOuputByOrderId}},
  \code{\link{getPipeline}}, \code{\link{getRegistry}},
  \code{\link{initialize}}, \code{\link{loadPipeline}},
  \code{\link{registerFunction}},
  \code{\link{savePipeline}}, \code{\link{updateObject}}
}
