% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark-utilities.R
\name{sparkRSessionCreateIfNotPresent}
\alias{sparkRSessionCreateIfNotPresent}
\title{Connect to a Spark session}
\usage{
sparkRSessionCreateIfNotPresent(...)
}
\details{
Loads the SparkR package and intializes a Spark session from R
}
\section{Slots}{

\describe{
\item{\code{...}}{Arguments to sparkR.session}
}}

\seealso{
Other Spark utilities: \code{\link{castKafkaStreamAsString}},
  \code{\link{convertKafkaValueFromJson}}
}
