% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-streaming-functions.R
\docType{class}
\name{StreamingAnalysisPipeline}
\alias{StreamingAnalysisPipeline}
\title{Class for constructing Analysis Pipelines for streaming analyeses}
\description{
Class for constructing Analysis Pipelines for streaming analyeses
}
\details{
Inherits the base class \link{BaseAnalysisPipeline} class which holds the metadata including the registry of available functions,
the data on which the pipeline is to be applied, as well as the pipeline itself

This class currently only supports Apache Spark Structured Streaming, implemented through the SparkR interface

More details of how an object of this class should be initialized is provided in the
constructor - \link{initializeStreamingAnalysisPipeline}
}
\section{Slots}{

\describe{
\item{\code{input}}{The input Spark DataFrame on which analysis is to be performed}

\item{\code{originalSchemaDf}}{Empty Spark DataFrame representing the schema of the input}
}}

\seealso{
Other Package core functions for Streaming Analyses: \code{\link{initializeStreamingAnalysisPipeline}}
}
\concept{Package core functions for Streaming Analyses}
