% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{updateObject}
\alias{updateObject}
\title{Update the \code{AnalysisRecipe} object by adding an operation to the recipe}
\usage{
updateObject(object, operation, heading = "", parameters, outAsIn = F)
}
\arguments{
\item{object}{object that contains input, recipe, registry and output}

\item{operation}{function name to be updated in tibble}

\item{heading}{heading of that section in report}

\item{parameters}{parameters passed to that function}

\item{outAsIn}{whether to use output of this function as input to next}
}
\value{
Updated \code{AnalysisRecipe} object
}
\details{
The specified operation along with the heading and parameters is updated in the recipe slot
      of the AnalysisRecipe object, where the sequence of operations to be performed is stored
}
\seealso{
Other Package core functions: \code{\link{generateOutput}},
  \code{\link{generateReport}}, \code{\link{initialize}},
  \code{\link{loadRecipe}}, \code{\link{readInput}},
  \code{\link{registerFunction}}, \code{\link{saveRecipe}},
  \code{\link{updatePackageRegistry}}
}
